Add function to backup dotfiles
This commit adds a new function, `backup_dotfiles`, which creates a tarball of the dotfiles and uploads it using rclone. If there is an error, a macOS notification is sent. The function also removes old log files and creates a symlink to the current log. Finally, it removes the tarball file.

The function is added in a new file named `backup_dotfiles.fish`.